+++ kompress-core/src/jsAndWasmMain/kotlin/dev/karmakrafts/kompress/lz77/LZ77.kt
@@ -16,7 +16,10 @@
 
 package dev.karmakrafts.kompress.lz77
 
-internal actual fun LZ77( // @formatter:off
+import dev.karmakrafts.kompress.InternalCompressionApi
+
+@InternalCompressionApi
+actual fun LZ77( // @formatter:off
     level: Int,
     minMatch: Int,
     maxMatch: Int,