+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## [Unreleased]
 
+### Added
+
+- ZLIB support via new `ZlibArchiver` and `ZlibUnarchiver` provided by `kompress-zlib`
+- `WrappingCompressor` delegate compressor to support wrapping existing compressors with extra data
+- `UnwrappingDecompressor` delegate decompressor to support unwrapping extra data with existing decompressors
+
+### Changed
+
+- Updated NMCP to 1.6.1
+
 ## [2.2.0]
 
 ### Added